Add basic file Structure
In this commit I added a basic file structure and boilerplate files.
The index.html has a link to the style.css file and the script.js file.

There is some test code in the style.css file to change the font colour and
background colour of the document to test it is linked correctly.

There is some test code in the script.js file to print out "testing" to
the console to check it is linked correctly.