Add basic html for the stack-section
This commit sets up the basic html markup for how the stacks and blocks
are to be structured.

The main purpose for this commit is to allow for initial styling of the
stacks and blocks and page so the main game development can start. This
structure is to be loaded dynamically by the javascript code and will be
removed once that feature is completed.