Add basic css styles to assist with development
This commit add in some basic styles to assist with development of the
game.

It adds it uses flex to align the action buttons in row and space them
out evenly.

It uses flex to align the stacks in a column and center them in the
middle of the page.

It sets some custom properties to allow maintain the dimensions of the
blocks as a square.

All these styles are likely to be adjusted once the game mechanics are
in place.