Add Game class with initial properties/methods
This commit adds in a Game class which will house all the logic for
starting a new Game object.

This commit sets 3 base properties and the associated methods.

There is a baseStackAmt property which is defaulted to 4 as this is the
minumum in which the game will be enjoyable.
There are associated get, set and update methods for baseStackAmt, with
the logic in the update method checking for incorrect data types if used
and not allowing a number below the default to be passed in.

There is a baseBlockAmt property which is defaulted to 4 as this is the
minumum in which the game will be enjoyable.
There are associated get, set and update methods for baseBlockAmt, with
the logic in the update method checking for incorrect data types if used
and not allowing a number below the default to be passed in.

There is a baseEmptyStackAmt property which is defaulted to 4 as this is the
minumum in which the game will be enjoyable.
There are associated get, set and update methods for baseEmptyStackAmt, with
the logic in the update method checking for incorrect data types if used
and not allowing a number below the default to be passed in.

This has been set up in this way to allow for any gameplay changes to be
made, as these values will drive the remaining game logic.