Add level, levelIncrements and startingStackAmt properties/methods
This commit adds the level, levelIncrements and startingStackAmt
properties and methods to the Game.

The level property is for the players current level and the methods
associated with it will get and set it, not allowing it to be set to
anything less than zero.

The levelIncrements property is an array of level increments in which
the game difficulty changes.

Both the level and levelIncrements properties are used in the
setStartingStackAmt method, in which it takes in the current level and
compares it against the levelIncrements array. The method then loops
through the length of the levelIncremenets array and checks the level
against the value in the array, if it returns true then it will take the
baseStackAmt and add on to it the index number of where it returned true
and increase the baseStackAmt.

There is a final else if that is there to ensure that once a certain
maximum amount of stacks are in play, the game will never increase past
this.