Add base colours property and methods
This commit adds the base colours to the game as a property of the Game
class, along with the associated methods.

In the interest of accessibility, I searched for colours that were unique enough,
but maintained a enough contrast/difference to be distinguishable. This lead me to
the following website with a great resource of colours trying to solve a similar problem
but for subway map lines. I selected a list of colours from this which I deemed
appropriate for the game.

 Link: https://sashamaps.net/docs/resources/20-colors/