Add inGameColours property and methods
This commit adds the inGameColours and methods to the Game class.

The below array will hold the in game colours. Essentially the method setInGameColours will
create a variable called coloursRequired and calculate the amount, based on the previously
created stacksToFill and baseBlockAmt properties.
Example:

stacksToFill = 2
baseBlockAmt = 4
coloursRequired = 2 * 4 = 8

It will then pull through the required amount of colours by referencing the index of the baseColours
array. Using the example numbers above, it will take the colour in the first index and push it to
a tempArray1 4 times, then it will move to the next index (less than stacksToFill) and push that colour
4 times to tempArray1.

The output of tempArray1 would look like this:

tempArray1 = ["red","red","red","red","green","green","green","green"]

There is a second part to this method, that randomises the colours in tempArray1 by creating a new tempArray2
variable and loops through up until the coloursRequired variable.
In the loop it will create a "multiplier" variable which represents the length of tempArray1 (note this array has items
spliced from it later). With this multiplier a random integer is found using Math.random. This is essentially an index in
the tempArray1 for a colour. That colour is then pushed on to tempArray2 and the colour is spliced from tempArray1.
This will result in a shuffled tempArray2 which is used to set the inGameColours property of the Game class.

The output of tempArray1 would look like this:

tempArray2 = ["green","green","red","red","red","green","red","green"]