Add Block and Stack classes and high level game flow comment
This commit adds two classes.
One for a Block, which will house the information for a Block object.
Another for a Stack, which will house the information for a Stack object
and the relevant methods to track inPlay variables about the Stack.

A high level description of how the game flow is going to work was added
as a comment above the Game class.