Add createBaseStacks method to Game class
This commit adds a gameStacks array property and a createGameStacks
method. This method uses the Block and Stack classes along with the
initialisation variables to create the correct amount of Stacks and
Blocks for the game.

Place holder methods to add colours to the blocks and id's to the
blocks and stacks also added.

Also added place holder methods to add the stacks to the DOM, as we will
be removing the hardcoded html currently there.

And two more methods, one to add the event listeners to the gameStack
elements that were placed in the DOM and the main game handler which
will be used to determine the first and second click while the game is
running.