Add setBlockColour method to Game class
This commit adds a method to add the colours to each of the blocks in
the gameStacks array. It loops through the defined properties set up
during the initialisation, utilising the stacksToFill property, the
baseStackAmt property and the inGameColours array property.

It adds the colours to the blocks sequentially then removes an element
from the inGameColours array. As the product of stacksToFill and
baseStackAmt equals the inGameColours, there will be no left over
elements in the inGameColours array once the loop has completed.

Visual checks were performed on the gameStacks object using chrome dev
tools to ensure that only 4 of each colour was populating and that they
were populating randomly (due to the randomness set in the
setInGameColours method).