Add the event listeners to the Stack elements
This commit adds the event listeners to the Stack elements in the DOM.

From there we can target the children in the Stack Node to check
properties of individual blocks and manipulate them for the game play.

Stack overflow and MDN was used to find a solution to an issue with the
"this" context when adding the event listeners. Both are documented
directly in the code and within the bugs section of the README.md,
explaining rationale and understanding of the usage of the arrow
function.

Placeholder console.logs were used for debugging this in the
handleGameClicks method.