Add targetting for the clicked stack node in the event handler
This commit adds targetting of the clicked stack node based on the click
event. It holds two main variables.

stackNode, which is the parent that was click and;

stackId, which is the id of the stack which will be used for targetting
both the DOM elements and the objects in the gameStack array throughout
the program.

A placeholder todo list of variables needed to get to the decision point
of determining the first click and second click has also been added.
Along with two Game class level properties of firstStackId and
secondStackId, which are important in the successful running of the
program.