Add methods to stack object to count full and empty stacks
This commit adds two methods as part of the Stack object.

One to counter the amount of full blocks in the Stack and one to count
the amount of empty blocks in the stack.

These functions will be used in conjuction with the next ones to
implement of isFilled and isEmpty, to help determine the click order of
the stack and prevent the clicking of the same stack on the second
click, and to prevent the clicking of an empty stack on the first click.