Add isFilled and isEmpty methods to Stack object
This commit adds two functions to assist with determining the first and
second click while playing the game.

They are isFilled and isEmpty and they check to see if the stack clicked
is either fully filled with colours or fully empty (void of coloured
blocks). The purpose is to enable tracking of if the first click is on
an empty stack (which is not possible) or if the second click is on the
a full stack (which is also not possible).