Add main handleGameClick logic to detect first/second clicks
This commit adds the functionality to detect the first and second game
clicks in order to set them as Game class level properties.

The functon is the main logic of the handleGameClicks method. Its purpose
is to determine which clicks are first and which are second. As the
Game class level properties are set as undefined by default, this if
statement first checks to see if the firstStackId is undefined, meaning
that it has not been set yet and it is therefore the first click. It will
then check to see if that stack is empty and if it is, the statement will return
due to not being able to click on an empty stack as the first click. At this
point the firstStackId and secondStackId are still undefined.

If it is not an empty stack it will set the firstStackId. When another click
occurs it will return false on the first check for undefined firstStackId because
it has already been set and move on to the else clause. Here in the else clause
it checks to see if the stack is filled (as you cannot click a filled stack for the
second click) or if the current event stackId is equal to the firstStackId
Game class level property set at the first click. If either returns true it
resets the firstStackId and SecondStackId Game level properties so the player
can start again. However if both of those return false it will continue to set
the secondStackId and call the compareBlocks method.