Add a clearGameStacks method and initialisation method
This commit adds two methods. One is a clearGameStacks method which
loops through all of the childnodes of the stack-section on the DOM and
removes each one until their are no childnodes remaining. This method
was from w3 schools and is attributed in the code.

The second method is a helper method to group all of the relevant
initialisation methods into one method. This allows the game to be set
up in less lines of code within the hasLoaded function. All that is
required to start a new game now is to define a Game object variable and
set it to new Game then to run the initialiseGame method on that new
game object.

Point of note, is that if you wanted to redefine the three base
properties of baseStackAmt, baseBlockAmt and baseEmptyStackAmt you would
do so after you have created the game object and just before you run the
initialiseGame method.