Interim jshint check
This commit is for the interim check of the code through jshint. I ran
it through jshint to mainly check for missing semi-colons, but it
highlighted that I had used a return on the createGameStacks method when
it wasnt needed. So I changed the line to update the Game class property
directly with the stacks array generated via the method.

The current results from jshint are as follows:
There are 59 functions in this file.
Function with the largest signature take 1 arguments, while the median is 0.
Largest function has 22 statements in it, while the median is 1.
The most complex function has a cyclomatic complexity value of 5 while the median is 1.

Using the following configuration

Cyclomatic complexity 			 	- on
Unused variables 			 	- on
Undefined variables 			 	- on

About == null				 	- on
About debugging code			 	- on
About unsafe for..in 			 	- on
About arguments.caller and .callee 	 	- on
About assignments if/for/... 		 	- on
About functions inside loops 		 	- on
About eval 				 	- on
About unsafe line breaks		 	- on
About potential typos in logical operators
When code is not in strict mode			- off
When new is used for side-effects		- on

Browser						- on
NodeJS						- off
jQuery						- off
Development (console, etc.)			- on
New JavaScript features (ES6)   		- on
Mozilla JavaScript extensions			- off
Older environments (ES3) 			- on