Add method to find index of top colour block on first click
This commit adds a method to find the index of the first block with a
colour for the first click. The property was added to the Stack class
and the method was called when the firstStackId is set during the
handleGameClicks method.

Will need to test more once there is the functionality to move blocks to
ensure it pulls the correct index.