Add two methods to compare blocks and move blocks
This commit adds two methods, one to compare the blocks to see if they
are eligible to be moved. First checking if the destination is empty and
moving straight away or checking if the destination top colour matches
the origin top colour then moving to the move blocks method.
If none of those are satisfied the stack id properties are reset on the
Game object.

The second method is a move blocks method, which in sync with each other
changes the DOM colours by changing the background colour of the
destination block to the origin block colour and resetting the origin to
blank. And then doing the same within the gameStacks array.

The end of the moveBlocks method always calls a hasWon method to check
if the game has been won after each move.