Add hasWon method to Game object
This commit adds the hasWon method to the Game object, along with a
helper method to check if the stacks are filled with the same colours.

The method starts with creating a variable called stackCompleteStatus
and a winningArray. There is then a loop which goes through each of the
stacks and calls the checkIfFilledWithSameColour method and returns true
or false, pushing that result into the winningArray.

There is then a check to see if the winningArray has the right amount of
true values. This is based on the amount of stacks currently minus the
baseEmptyStackAmt variable.
For example if there were a total of 4 stacks and 2 empty then the game
will be classed as won if the winningArray was [true, false, true,
false].

The initialiseGame method is then called if the game is won to start a
new game.