Add move logging functionality and placeholder event listeners for btns
This commit adds in a few lines of code to the moveBlocks method to
track the moves that have been made. This is so that the functionality
for the undoMove and resetLevel method can reference it and change the
state of the game. Placeholder functions have been added for 3 methods
along with the event listeners.
- undoMove
- resetLevel
- addBlock