Add undoMove and resetLevel methods
This commit has two methods added to the event listeners to undoMove and
resetLevel.

undoMove reads the information set in the moves array property and
reverses the previous move then pops that move off the array.

resetLevel is just an iteration of undoMove for the entire length of the
moves array.