Add the functionality to add bonus block
This commit is quite a comprehensive change as it changes differnt parts
of the code base to incorporate the functionality to add a supporting
block to the game to allow for game completion.

Firstly it adds 3 new properties to the Stack object.

It add a maxBlockAmt property. This property is set as part of the
createGameStacks method within the Game class object. Its purpose is to
accurately show how many baseBlocks should be part of the initial stack
object. It is used in conjunction with the next property which is
bonusBlock.

bonusStack is a property that is set to false by default on the creation
of the gameStacks and is only set to true when the addBlock method is
called. Its purpose is to be able to distinguish from a "normal" stack
and a "bonus" stack that was added as part of buying a supporting block.
It is directly used within the hasWon method for determining which
stacks it needs to check are full. (more information below on the
changes to the hasWon method).

The third property added is the updateIsFilledWithSameColour, which is a
refactor of the previous Game object level method, but placed now within
the Stack class. This gets called still during the hasWon method to
determine if a stack is complete.

The hasWon method was refactored to reference the gameStacks object
instead of the DOM, so to have access to all of the supporting methods
as part of the Stack object. It maintains a similar structure in which
it creates a winningArray, but now this is based on the stacks that are
set to bonusBlock = false. This is to ensure that the non-bonus stacks
are the only ones being checked for "fullness of same colour".

And the final method added, which was the main point of this update is
the addBlock. All of the above supporting methods and properties now
allow for this method to not impact the game mechanics already designed.

The current solution for addBlock creates a new stack and gives it a
true value on the bonusStack property along with relevant ids and
classes to it can still utilise the referencing used in the moveBlocks
method. Once a stack is created it is added to the gameStacks object
then the unedited and existing method of clearGameStacks, addStackToDOM
and addEventListenersStackArea are called, in order to "re-draw" the
gameStacks object to the DOM.

There is scope to update the addBlock function to append new blocks to
an existing bonusStack, but in its current form it only creates a new
stack each time with one block in it.

Testing was completed with the game up to the highest level possible,
and all expected gameplay functionality exists.