Add additional addBlock functionality
This commit adds additional functionality to add more than 1 bonus block
to the game. Currently set at a hard limit of 2 per level.

It enhances the previous code by appending a block to the current bonus
stack and making sure the colours shift accordingly.

If there is a requirement to add more than 2 blocks, this code will need
to be amended to account for the shifting of the colours. However I felt
there wasnt a need to have more than 2 bonus blocks per level so it was
not factored into the code.

Below is the full comment placed above the method

        The below method allows for a bonus block to be added to the game to assist the
        player in completing a level.
        It first checks to see if the player already has the maximum amount of blocks, if
        they already have the maximum amount the method will return with no action.
        TODO - potentially add some feedback that they have used the max amount (update of
        a number in the button or a notification).

        If the player is eligible to add a bonus block, the method will first check to see if
        there is already a bonus stack in play, if there isnt then it will create a stack along
        with a block. If there is a stack already in play it will only add a new block to the
        existing bonus stack.

        Once the block/stack has been added, if there was a colour in the first bonus block, this
        is shifted down to the bottom block in the stack.

        Finally the method calls 3 existing methods to clear the stacks, re-add the stacks to the DOM
        and then re-add the event listeners.