Adjust levelIncrements and colour array and order
This commit adjusts the colour array by reducing the amount of colours
available and moving up the more distinguishable colours to be part of
the game colours.

Also added an improved level increment array. I used a factor of 1.55 to
multiply each previous level increment to make the increases feel
staggered and natural.