Added styling to title, icons and background of game
This commit adds a colourful styling to the main header using css to
define each individual letter colour.

It also adds new icons for the reset, undo and add block buttons, which
also scale in size the same as the block buttons.

Some minor adjustments to the spacing and letter sizing across the game.