Add animation to game title
This commit adds some simple animation to the game title for each time
a new level starts.

There is some unused placeholder code within the GameTitle class at the
moment that is for a future feature. In which the logic of the code will
check to see if there colour of the first letter of Colour and first
letter of Sort are the same and if so there will be a bonus round.