Add bonus level functionality
This commit adds a bonus level functionality, where it hides the colours
on the DOM from view until a block has been moved.

Created a new method to set a boolean value to true if the level is
modulus = 0 to a certain number (2 for testing). Then it will call an
additional method during the initialisation stage to hide the colours on
the DOM.

There is then a check within the moveBlocks method to check for this
boolean value and unhide the colours as the blocks move. There is also a
check within the reset level to make sure that the bonus level blocks
get hidden properly again.