Add LevelManager class
This commit adds a LevelManager class to return the starting stack
amount needed for each level based on the level increments.

The LevelManager class has one method which takes in the current level
from the GameManager properties and the base stack amount from the
GameSettings class. The method has error checking to assist the
developer if they pass in an incorrect parameter type.