Add ColourInitialiser class
This commit adds a new class called ColourInitialiser. It is a refactor
of a few methods that were part of the original Game class.

The change with the method vs the previous method is that now it
generates a full array of items, both colours needed and undefined
values for none coloured blocks. Which makes it easier to assign the
colours for the game.

The method also contains error checking to assist the developer, by
throwing errors if incorrect data types are passed as parameters.

The functionality in general remains the same. It takes in the
parameters of how many stacks the game needs, how many blocks per stack
and how many empty stacks there should be.

It creates an array of colours and randomly selects them to create the
starting colours then appends to that array the correct amount of
undefined colour values for the block.