Add GameStacks class
This commit add the GameStacks class with a method to create the initial
game stacks.

It takes in the parameters of how many stacks needs to be created and
the amount of blocks per stack. Then calls the Stack class and Block
class to create those objects and store them in a gameStacks property on
the GameManager.