Add methods to Stack to check for state
This commit adds two methods to the Stack class to check for the state
of if the stack is empty or if the stack is filled.

It is a refactor of the previous code in that it updates all stacks at
once for the current state of them by running a loop inside the event
handler. Also utilising the map and every array methods.
The map method pulls out the colour from the array of blocks in the
Stack and then the every method checks to see if they are either all
undefined or either all not undefined, to set the isEmpty and isFilled
properties on the Stack class respectively.