Add two more methods to Stack to check for state
This commit adds two more methods to the Stack class to check for state
of the stack. It refactors the old code by consolidating the first click
and second click methods into methods that can check the state of the
stack regardless of what click it is. It also removes the methods that
originally checked for index values and utilises the findIndex array
method.