Add logic for first/second click
This commit adds the logic to the handleGameClicks method to determine
the first and second click. It remains fairly unchanged from the
previous code. Apart from now it does not have the compareBlocks method
associated with the Game class. That has been encapsulated within the
GameStacks class within a method called getNewGameStacksState. From this
method, the comparing of Blocks and moving of Blocks will occur within
the GameStacks object, and then be returned back to the GameManager to
be drawn back to the DOM for any changes.