Add logic to getNewGameStacksState method on GameStacks class
This commit adds the logic to the getNewGameStacksState method. It is a
consolidation and refactor of the old compareMoves and moveBlocks
methods, which can now be condensed due to the simlification of the
state properties on the Stack class and also the separation of concerns
between the gameStacks updating and the DOM drawing.