Add hasWon method to GameStacks
This commit adds the hasWon method to the GameStacks class. The hasWon
method was originally on the Game class but it has now been encapsulated
within the GameStacks object. It will now return a boolean value back to
the handleGameClicks handler and if there is a win state it will trigger
a new game.

This commit also adds a support state method to the stack object. In
which it checks if the stacks are filled with the same colour, and this
is used to check for win conditions.