Add functionality to addBlock to gameStacks
This commit adds the functionality to add a support block to help solve
the game. Works for a maximum of 2 blocks.

It adds the additional Stack and Block to the stack if there isnt
already a block. Then if a second one is added it checks if a stack
already exists and just appends the new block. Moving the colours into
the right position.
This method then returns the gameStacks back to the GameManager, where
it can be redrawn to the DOM.