Add undoMove method to GameStacks class
This commit adds the undoMove method to the GameStacks class. It is
possible due to the moves being logged as part of the
getNewGameStacksState method storing the moves in a property on the
GameStacks class in movelog.

The undoMove method checks this moveLog and reverses the last move.
Returning the GameStacks object back to the GameManager so it can be
redrawn on the DOM.

It links up the undoMove button with an event handler in GameManager
which calls the GameStacks undoMove method then updates the DOM.