Add resetLevel method to GameStacks class
This commit adds the resetLevel method to the GameStacks class which
recursivly calls the undoMove method until there are no more elements in
the moveLog array.

It links up the buttons to their event handlers, which calls the
resetLevel method on the GameStacks, then redraws the relevant elements
and event handlers to the DOM for the game to continue.